+++ build_cli/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Fail unless the minimum SDK constraint on the target package is at least
   `2.0.0-dev.48`.
 
+- Added support `valueHelp` to `CliOption`.
+
 ## 0.1.2
 
 - Support `int`, `double`, and `num` for options.